double linked moved to ball
